--- kernel/src/checkpoints/mod.rs
@@ -1,1 +1,0 @@
-pub mod log_replay;
